refactor(workhub): retire legacy route policy seam
Generated-by: Codex
